requirements.txt
meteostat
python-telegram-bot==13.15
pandas
matplotlib
seaborn
requests